Builds with VS2015.

+++ Source/Common/Include/latticearchive.h
@@ -874,7 +874,7 @@ public:
     template <typename HMMLOOKUPFUNCTION>
     void dump(FILE* f, const HMMLOOKUPFUNCTION& gethmmname) const // dump a lattice in HTK-like format
     {
-        fprintf(f, "N=%lu L=%lu\n", nodes.size(), edges.size());
+        fprintf(f, "N=%zu L=%zu\n", nodes.size(), edges.size());
         // foreach_index (i, nodes)
         //    fprintf (f, "I=%d\tt=%.2f\n", i, nodes[i].t * 0.01f);
         foreach_index (j, edges)
@@ -1281,7 +1281,7 @@ public:
             char c;
             uint64_t offset;
 #ifdef _WIN32
-            if (sscanf_s(q, "[%I64u]%c", &offset, &c, sizeof(c)) != 1)
+            if (sscanf_s(q, "[%I64u]%c", &offset, &c, (unsigned int)sizeof(c)) != 1)
 #else
 
             if (sscanf(q, "[%" PRIu64 "]%c", &offset, &c) != 1)
